requirements.txt
discord.py
mcstatus
paramiko
pytz
python-dotenv